+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/types/RealmLinkingObjects.kt
@@ -16,19 +16,8 @@
 
 package io.realm.kotlin.types
 
-import io.realm.kotlin.internal.RealmLinkingObjectsImpl
 import io.realm.kotlin.query.RealmResults
-import kotlin.reflect.KClass
 import kotlin.reflect.KProperty
-import kotlin.reflect.KProperty1
-
-// TODO document
-public fun <T : RealmObject> linkingObjects(targetProperty: KProperty1<T, *>, clazz: KClass<T>): RealmLinkingObjects<T> =
-    RealmLinkingObjectsImpl(targetProperty, clazz)
-
-// TODO document
-public inline fun <reified T : RealmObject> linkingObjects(targetProperty: KProperty1<T, *>): RealmLinkingObjects<T> =
-    linkingObjects(targetProperty, T::class)
 
 // TODO document
 public interface RealmLinkingObjects<T : RealmObject> {
